requirements.txt
asgiref==3.5.2
Django==4.1.2
django-ckeditor==6.5.1
django-js-asset==2.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Pillow==9.3.0
Pygments==2.13.0
PyJWT==2.6.0
pytz==2022.6
sqlparse==0.4.3
tzdata==2022.5
